CMake improvements.
Increased CMAKE_CXX_STANDARD to 14 to prepare for GTest dropping C++11 for
future releases. This should still fall back to C++11 for older targets.

Removed MACOSX_RPATH since it was incorrectly named (should start with
CMAKE_) and is the default starting with CMake 3.0 and later.

Avoid installing package config if install is disabled.